refactor: split CLI into modules, remove mcp-server
