+++ hotspots-cli/src/cmd/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod analyze;
+pub(crate) mod compact;
+pub(crate) mod config;
+pub(crate) mod prune;
+pub(crate) mod trends;
