--- packages/mcp-server/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@hotspots/mcp-server",
-  "version": "1.0.0",
-  "description": "Model Context Protocol server for Hotspots complexity analysis",
-  "main": "dist/index.js",
-  "bin": {
-    "hotspots-mcp": "dist/index.js"
-  },
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "start": "node dist/index.js"
-  },
-  "keywords": [
-    "mcp",
-    "hotspots",
-    "claude",
-    "complexity",
-    "analysis",
-    "model-context-protocol"
-  ],
-  "author": "Stephen Collins",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Stephen-Collins-tech/hotspots.git",
-    "directory": "packages/mcp-server"
-  },
-  "homepage": "https://hotspots.dev",
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "^0.5.0",
-    "execa": "^8.0.1"
-  },
-  "devDependencies": {
-    "@types/node": "^20.0.0",
-    "typescript": "^5.3.0"
-  }
-}
